Fixed median() bad math, made binary search available in remove()

+++ Stats.h
@@ -22,7 +22,7 @@ class Stats
 		int maximum();
 		int sum();
 		double mean();
-		int median();
+		double median();
 		void display();
 		
 		// Manipulate collection, unsanitary class methods.
